Convert osu match container to act as single server Fixes #1
