+++ Dockerfile
@@ -36,4 +36,4 @@ COPY --from=prerelease /usr/src/app/.env .
 # run the app
 USER bun
 EXPOSE 3000/tcp
-ENTRYPOINT [ "bun", "run", "index.ts" ]
+ENTRYPOINT [ "bun", "run", "src/index.ts" ]
