+++ package.json
@@ -3,7 +3,8 @@
   "module": "index.ts",
   "type": "module",
   "devDependencies": {
-    "@types/bun": "latest"
+    "@types/bun": "latest",
+    "@types/express": "^4.17.21"
   },
   "peerDependencies": {
     "typescript": "^5.0.0"
@@ -11,6 +12,7 @@
   "dependencies": {
     "@supabase/supabase-js": "^2.43.5",
     "bancho.js": "^0.11.6",
+    "express": "^4.19.2",
     "osu-api-extended": "beta"
   }
 }
