+++ src/endpoints/sendMessages.ts
@@ -1,0 +1,3 @@
+export async function sendMessages(messages: string[]) {
+  console.log(req.body);
+}
