+++ src/events/playerMoved.ts
@@ -1,0 +1,1 @@
+export async function playerMoved() {}
