+++ src/events/playerReady.ts
@@ -1,0 +1,7 @@
+export async function playerReady() {
+  channel.lobby.on("allPlayersReady", async () => {
+    console.log("All players are ready");
+
+    channel.lobby.startMatch();
+  });
+}
