+++ src/index.ts
@@ -1,0 +1,31 @@
+import { BanchoClient } from "bancho.js";
+import express from "express";
+import type { Request, Response } from "express";
+import { createMatch } from "./endpoints/createMatch.ts";
+import { sendMessages } from "./endpoints/sendMessages.ts";
+
+console.log("Starting osu! server.");
+
+const banchoClient = new BanchoClient({
+  username: process.env.OSU_USERNAME!,
+  password: process.env.OSU_IRC_KEY!,
+  apiKey: process.env.OSU_API_KEY!,
+});
+await banchoClient.connect();
+
+console.log("Connected to Bancho");
+
+const app = express();
+const port = 3000;
+app.use(express.json());
+
+app.post("/create-match", async (req: Request, _res: Response) => {
+  await createMatch(req.body.id, banchoClient);
+});
+app.post("/send-message", async (req: Request, _res: Response) => {
+  await sendMessages(req.body.messages);
+});
+
+app.listen(port, () => {
+  console.log(`Server running on port ${port}`);
+});
